+++ wallet_data/src/commonMain/kotlin/net/thechance/mena/wallet/data/dto/remote/TransactionDto.kt
@@ -18,5 +18,5 @@ data class TransactionDto(
     @SerialName("createdAt")
     val createdAt: String? = null,
     @SerialName("amount")
-    val silvers: Double? = null
+    val amount: Double? = null
 )
