+++ wallet_data/src/commonMain/kotlin/net/thechance/mena/wallet/data/mapper/TransactionMapper.kt
@@ -1,7 +1,6 @@
 package net.thechance.mena.wallet.data.mapper
 
 import net.thechance.mena.wallet.data.dto.remote.TransactionDto
-import net.thechance.mena.wallet.domain.entity.Currency
 import net.thechance.mena.wallet.domain.entity.Transaction
 import net.thechance.mena.wallet.domain.entity.TransactionStatus
 import net.thechance.mena.wallet.domain.entity.TransactionType
@@ -20,7 +19,7 @@ fun TransactionDto.toEntity(): Transaction {
         status = TransactionStatus.valueOfOrDefault(status),
         senderName = senderName ?: "",
         receiverName = receiverName ?: "",
-        amount = Currency(silvers = silvers ?: 0.0),
+        amount = amount ?: 0.0,
         type = TransactionType.valueOfOrDefault(type),
     )
 }
