--- wallet_domain/src/commonMain/kotlin/net/thechance/mena/wallet/domain/entity/Currency.kt
@@ -1,5 +1,0 @@
-package net.thechance.mena.wallet.domain.entity
-
-data class Currency(
-    val silvers: Double
-)
