--- wallet_presentation/src/commonMain/kotlin/net/thechance/mena/wallet/presentation/model/ToastState.kt
@@ -1,8 +1,0 @@
-package net.thechance.mena.wallet.presentation.model
-
-import org.jetbrains.compose.resources.StringResource
-
-data class ToastState(
-    val isVisible: Boolean = false,
-    val messageRes: StringResource? = null,
-)
